+++ src/ic-cdk/CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+
+- `offset` for exposing offsets of `StableReader` and `StableWriter` implementations
+
 ## [0.6.3] - 2022-10-26
 
 ### Fixed